Better dependency handling
