+++ .gitignore
@@ -55,3 +55,4 @@ dkms.conf
 .idea
 build
 .DS_Store
+.vscode
