+++ .travis.yml
@@ -11,7 +11,7 @@ before_install:
     - make
     - sudo make install
     - cd ../../
-    -
+    
 compiler:
     - clang
     - gcc
@@ -21,3 +21,19 @@ script:
     - cd build 
     - cmake .. 
     - make
+
+before_deploy:
+    - mkdir export
+    - mv uspm* export/
+    - cd export
+    - tar cf uspm-linux.tar ./*
+    - mv uspm-linux.tar ../../
+    - cd ../../
+
+deploy:
+    provider: releases
+    api_key: $ghtoken
+    file: "uspm-linux.tar"
+    skip_cleanup: true
+    on:
+        tags: true
