+++ TODO.md
@@ -1,7 +1,8 @@
 # TODO
 
 ## Version 1.1
-1) Install multiple packages all at the same time
-2) When installing a dependency, check to make sure its version is above minimum version
+~~1) Install multiple packages all at the same time~~
+1) When installing a dependency, check to make sure its version is above minimum version
+
 
 # Next version
